+++ app/constants.js
@@ -1,0 +1,2 @@
+module.exports.NAME_DELIMITER = "|";
+module.exports.AUTH_CHALLENGE_TIME_TO_LIVE = 60; //seconds. Forward or backward from server time
